+++ stream-chat-android-client/src/test/java/io/getstream/chat/android/client/ConnectUserTest.kt
@@ -21,8 +21,9 @@ internal class ConnectUserTest {
         socketStateService = mock(),
         queryChannelsPostponeHelper = mock(),
         userCredentialStorage = mock(),
-        appContext = mock(),
         scope = mock(),
+        retryPolicy = mock(),
+        errorHandlerFactory = mock(),
     )
 
     @Test