+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/message/list/adapter/MessageListItemViewHolderFactory.kt
@@ -12,6 +12,7 @@ import io.getstream.chat.android.ui.message.list.adapter.MessageListItemViewType
 import io.getstream.chat.android.ui.message.list.adapter.MessageListItemViewType.FILE_ATTACHMENTS
 import io.getstream.chat.android.ui.message.list.adapter.MessageListItemViewType.GIPHY
 import io.getstream.chat.android.ui.message.list.adapter.MessageListItemViewType.GIPHY_ATTACHMENT
+import io.getstream.chat.android.ui.message.list.adapter.MessageListItemViewType.IMAGE_ATTACHMENT
 import io.getstream.chat.android.ui.message.list.adapter.MessageListItemViewType.LOADING_INDICATOR
 import io.getstream.chat.android.ui.message.list.adapter.MessageListItemViewType.MESSAGE_DELETED
 import io.getstream.chat.android.ui.message.list.adapter.MessageListItemViewType.PLAIN_TEXT
@@ -28,6 +29,7 @@ import io.getstream.chat.android.ui.message.list.adapter.viewholder.internal.Err
 import io.getstream.chat.android.ui.message.list.adapter.viewholder.internal.FileAttachmentsViewHolder
 import io.getstream.chat.android.ui.message.list.adapter.viewholder.internal.GiphyAttachmentViewHolder
 import io.getstream.chat.android.ui.message.list.adapter.viewholder.internal.GiphyViewHolder
+import io.getstream.chat.android.ui.message.list.adapter.viewholder.internal.ImageAttachmentViewHolder
 import io.getstream.chat.android.ui.message.list.adapter.viewholder.internal.MessageDeletedViewHolder
 import io.getstream.chat.android.ui.message.list.adapter.viewholder.internal.MessagePlainTextViewHolder
 import io.getstream.chat.android.ui.message.list.adapter.viewholder.internal.SystemMessageViewHolder
@@ -132,6 +134,7 @@ public open class MessageListItemViewHolderFactory {
             THREAD_PLACEHOLDER -> createEmptyMessageItemViewHolder(parentView)
             GIPHY_ATTACHMENT -> createGiphyAttachmentViewHolder(parentView)
             FILE_ATTACHMENTS -> createFileAttachmentsViewHolder(parentView)
+            IMAGE_ATTACHMENT -> createImageAttachmentViewHolder(parentView)
             else -> throw IllegalArgumentException("Unhandled MessageList view type: $viewType")
         }
     }
@@ -160,9 +163,32 @@ public open class MessageListItemViewHolderFactory {
      * @return The [BaseMessageItemViewHolder] that holds a Giphy image in various quality types.
      */
     protected fun createGiphyAttachmentViewHolder(
-        parentView: ViewGroup
+        parentView: ViewGroup,
     ): BaseMessageItemViewHolder<out MessageListItem> {
-        return GiphyAttachmentViewHolder(parentView, decoratorProvider.decorators, listenerContainer, markdown = textTransformer)
+        return GiphyAttachmentViewHolder(
+            parentView,
+            decoratorProvider.decorators,
+            listenerContainer,
+            markdown = textTransformer
+        )
+    }
+
+    /**
+     * Creates a ViewHolder for messages containing image attachments and no other type
+     * of attachments.
+     *
+     * @param parentView The parent container.
+     * @return The [BaseMessageItemViewHolder] that displays messages with image attachments.
+     */
+    protected fun createImageAttachmentViewHolder(
+        parentView: ViewGroup,
+    ): BaseMessageItemViewHolder<out MessageListItem> {
+        return ImageAttachmentViewHolder(
+            parentView,
+            decoratorProvider.decorators,
+            listenerContainer,
+            textTransformer,
+        )
     }
 
     /**