+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/message/list/adapter/viewholder/decorator/internal/BaseDecorator.kt
@@ -7,6 +7,7 @@ import io.getstream.chat.android.ui.message.list.adapter.viewholder.internal.Dat
 import io.getstream.chat.android.ui.message.list.adapter.viewholder.internal.FileAttachmentsViewHolder
 import io.getstream.chat.android.ui.message.list.adapter.viewholder.internal.GiphyAttachmentViewHolder
 import io.getstream.chat.android.ui.message.list.adapter.viewholder.internal.GiphyViewHolder
+import io.getstream.chat.android.ui.message.list.adapter.viewholder.internal.ImageAttachmentViewHolder
 import io.getstream.chat.android.ui.message.list.adapter.viewholder.internal.MessageDeletedViewHolder
 import io.getstream.chat.android.ui.message.list.adapter.viewholder.internal.MessagePlainTextViewHolder
 import io.getstream.chat.android.ui.message.list.adapter.viewholder.internal.TextAndAttachmentsViewHolder
@@ -27,6 +28,7 @@ internal abstract class BaseDecorator : Decorator {
             is GiphyViewHolder -> decorateGiphyMessage(viewHolder, data)
             is GiphyAttachmentViewHolder -> decorateGiphyAttachmentMessage(viewHolder, data)
             is FileAttachmentsViewHolder -> decorateFileAttachmentsMessage(viewHolder, data)
+            is ImageAttachmentViewHolder -> decorateImageAttachmentMessage(viewHolder, data)
             is DateDividerViewHolder -> Unit
             else -> Unit
         }.exhaustive
@@ -56,6 +58,17 @@ internal abstract class BaseDecorator : Decorator {
      */
     abstract fun decorateFileAttachmentsMessage(
         viewHolder: FileAttachmentsViewHolder,
+        data: MessageListItem.MessageItem,
+    )
+
+    /**
+     * Applies various decorations to the [ImageAttachmentViewHolder].
+     *
+     * @param viewHolder The holder to be decorated.
+     * @param data The data used to define various decorations.
+     */
+    abstract fun decorateImageAttachmentMessage(
+        viewHolder: ImageAttachmentViewHolder,
         data: MessageListItem.MessageItem,
     )