+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/message/list/background/MessageBackgroundFactoryImpl.kt
@@ -30,6 +30,16 @@ public open class MessageBackgroundFactoryImpl(private val style: MessageListIte
     }
 
     /**
+     * Draws the background of messages containing image attachments.
+     *
+     * @param context [Context].
+     * @param data [MessageListItem.MessageItem].
+     */
+    override fun imageAttachmentMessageBackground(context: Context, data: MessageListItem.MessageItem): Drawable {
+        return defaultBackground(context, data)
+    }
+
+    /**
      * Draws the background of deleted messages,
      *
      * @param context [Context].